prevent split packages
