+++ services/infinispan-device-registry/src/main/java/org/eclipse/hono/registry/infinispan/CacheRegistrationConfigProperties.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package org.eclipse.hono.deviceregistry;
+package org.eclipse.hono.registry.infinispan;
 
 /**
  * Configuration properties for Hono's Infinispan Registration service.
