+++ services/infinispan-device-registry/src/main/java/org/eclipse/hono/registry/infinispan/CacheRegistrationService.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package org.eclipse.hono.deviceregistry;
+package org.eclipse.hono.registry.infinispan;
 
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Future;
