+++ services/infinispan-device-registry/src/main/java/org/eclipse/hono/registry/infinispan/CacheTenantConfigProperties.java
@@ -10,11 +10,11 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package org.eclipse.hono.deviceregistry;
+package org.eclipse.hono.registry.infinispan;
 
 /**
- * Configuration properties for Hono's Infinispan Credential service.
+ * Configuration properties for Hono's Infinispan Tenant service.
  *
  */
-public class CacheCredentialConfigProperties {
+public class CacheTenantConfigProperties {
 }
